Fix retire fences
The HWC2 header does not require deferring the retire fence, also
known as the present fence. The retire fence is now known as the
present fence in the HWC2 header and it belongs to the
current frame that is submitted and is signalled when the current
frame is picked up. Also, the retire fence was incorrectly set to
the release fence of the output buffer. Use the retire fence
instead.

Bug: 30204952
Change-Id: Ide2fdcc27d0ed42119ccb1e37e16d43c32a39790